vision-engine/requirements.txt
# Vision engine dependencies for Actual Reader
# Qwen2.5-VL-7B for image captioning

transformers>=4.45.0
accelerate>=0.30.0
qwen-vl-utils
pillow>=10.0.0

# PyTorch - shared with TTS engine
# torch>=2.2.0 (install separately based on CUDA/ROCm/MPS)
